// app/src/api/vault/VaultClient.ts
/** @generated by @calimero/abi-codegen — DO NOT EDIT. */

import {
  MeroJs,
} from '@calimero-network/mero-react';

// Generated types

export interface Entry {
  data: EntryData;
  created_at: number;
}

export interface EntryData {
  service_name: string;
  username: string;
  secret: string;
  notes: string;
}

export interface Event_EntryAdded {
  id: string;
  author: string;
}

export interface Event_EntryDeleted {
  id: string;
}

export interface Event_EntryUpdated {
  id: string;
}

export interface Vault {
  entries: Record<string, Entry>;
}

export interface VaultEntry {
  id: string;
  author: string;
  service_name: string;
  username: string;
  secret: string;
  notes: string;
  created_at: number;
}




export type AbiEvent =
  | { name: "EntryAdded"; payload: Event_EntryAdded }
  | { name: "EntryUpdated"; payload: Event_EntryUpdated }
  | { name: "EntryDeleted"; payload: Event_EntryDeleted }
;


export class VaultClient {
  private _mero: MeroJs;
  private _contextId: string;
  private _executorPublicKey: string;

  constructor(mero: MeroJs, contextId: string, executorPublicKey: string) {
    this._mero = mero;
    this._contextId = contextId;
    this._executorPublicKey = executorPublicKey;
  }

  /**
   * init
   */
  public async init(): Promise<void> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'init', argsJson: {}, executorPublicKey: this._executorPublicKey });
    return response as void;
  }

  /**
   * add_entry
   */
  public async addEntry(params: { service_name: string; username: string; secret: string; notes: string }): Promise<string> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'add_entry', argsJson: params, executorPublicKey: this._executorPublicKey });
    return response as string;
  }

  /**
   * edit_entry
   */
  public async editEntry(params: { id: string; service_name: string; username: string; secret: string; notes: string }): Promise<void> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'edit_entry', argsJson: params, executorPublicKey: this._executorPublicKey });
    return response as void;
  }

  /**
   * delete_entry
   */
  public async deleteEntry(params: { id: string }): Promise<void> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'delete_entry', argsJson: params, executorPublicKey: this._executorPublicKey });
    return response as void;
  }

  /**
   * list_entries
   */
  public async listEntries(): Promise<VaultEntry[]> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'list_entries', argsJson: {}, executorPublicKey: this._executorPublicKey });
    return response as VaultEntry[];
  }

}